Had to define a global variable cl_do_doinp (sigh).  This is set to True (1) by default.  Only in loadDefaults() it is set to False (0) at the beginning and set to True (1) before returning.  This ensures that the call to SetVar from loadDefaults() does not call doinp() to display variables.  From everywhere else, SetVar() calls doinp to display any newly exposed parameters.
This also means that newly exposed parameters due to doload() will not
be displayed.  This might be a desiable thing to do to keep the
behaviour consistant with "newly displayed parameters will be shown
when they are exposed due to new settings".  But this can wait for
another bug fix and checkin :).