Changes for GCC 4.4.x (on RHEL6)
Removed a bunch of warnings by type-casting strings with (char *).

Using %option noyywrap in shell.l to remove dependence on yywrap()
which the complier does not seem to fine in libfl.a.

Fixed the stripwhite() function in support.c.  The source and target
strings in strcpy() were the same and does not give correct results
any more with GCC 4.4.x.  Using std::string as a temporay storage
now -- which means that this has to be necessarily compiled with G++.