Using the clget[N]Valp() call in clget?Valp() wrappers.
This makes the clget[N]Valp() kind of template call a step away from
being an API-level call.  The clget?Valp() and clgetN?Valp() calls are
for backward compatibility (or when compiled with strict pure-C98
compiler).

clget[N]?Valp() calls are not yet usable as API calls.  That's coming
next.

Added the "run" and "runb" interactive shell commands.