Added help=params option.
With this option applications will print a table of information about
the application parameter names, type, default value and options (if
they exists).  This could be used to ensure all (?) interfaces present
similar interface.

An example of the table is below.  Lines starting with %%N gives the
name of the application.  Each line starting with %%P has the
following format:

Name : type : default value : options

>tstcpp help=params
%%N tstcpp
%%P bool : bool : 0 :
%%P bool1 : bool : 0 :
%%P int : int : 0 :
%%P dbgint : int : 1 :
%%P float : float : 0 :
%%P oneint : int : 0 :
%%P string : string : showstrarr : one|two|three|showstrarr|testarr|
%%P strarr : string[] :  :
%%P farray : float[3] : 0,0,0,0,0,0,0,0,0,0 :