Convert symbols in cl_SymbTab to CL_USERCLASS after load the full .def file in doload_and_register().
This allows setting the value of the last occurrance of a symbol.
Symbols that are queried vis clget*Val() in the application code get
converted to CL_APPLNCLASS and only those are then printed.  This
allows use of a .def to all apps, which may be a combination of
parameters of multiple apps in LibRA.  Each app will then print only the
keywords it uses.