Mostly changes (cleanup + corrections) in clgtValp.{cc,h}
Functions are consolidated into fewer templates.
clgetValp.cc is not included in clgetValp.h.
clgetValp.cc has to be included in applications that use the clgetValp() interface.
It therefore also has to be installed.
clget*Valp() interfaces are implemented in clgetValp.cc.  So it has to be compiled into the library.
Using code/test/tst{2,3}.cc programs more better testing.